+++ concerns/data_types/words.c
@@ -3,8 +3,8 @@
 #include "vector.h"
 
 char *first_word(words words, char *line) {
-  size_t a = (size_t)vector_get(words, 0);
-  size_t b = (size_t)vector_get(words, 1);
+  unsigned a = (unsigned)vector_get(words, 0);
+  unsigned b = (unsigned)vector_get(words, 1);
   char *s = strdup(line + a);
   s[b - a + 1] = '\0';
   return s;
